Now added the ability to have the Perfmon4j instrumentation agent disable the method System.gc(). This is useful in place of the Sun JVM parameter "-XX:+DisableExplicitGC".
The problem with the Sun parameter is that it removes any option to force a heap compaction if you are using the concurrent collector.  This means that using the concurrent collector will eventually lead to a fragmented heap.  Disabling System.gc() still leaves open the ability to call Runtime.getRuntime().gc() at off peak intervals to force a compaction.

git-svn-id: http://svn.code.sf.net/p/perfmon4j/code/trunk@114 38ac2094-77c1-42ba-995b-c18f1c0364f1