Replaced high throughput synchronized blocks with explicit Reentrant locks.  There was some evidence that the synchronized blocks were not performing optimally in high throughput/highly multi-threaded environments.
Also fixed a defect with the new dynamically generated classes that hold monitor references when instrumenting serialized classes that cannot have their signature modified.  The problem was uncovered during regression testing with the WildFly application server and was based on the dynamic class not being associated with the same class loader as the instrumented class.

git-svn-id: http://svn.code.sf.net/p/perfmon4j/code/trunk@376 38ac2094-77c1-42ba-995b-c18f1c0364f1