Found the embedded javassist classloader does not work under apache tomcat.  Most likely based on a security policy.
It does however work under wildfly and through all of the unit tests.
Now we first check the embedded classloader to validate we can load a
javassist class.  If we cant (likem under tomcat) we fall back and
require the external javassist.jar.

This means that for a tomcat install we will still require javassist.jar
is copied to the tomcat/lib/endorsed folder.