Support environment variables in config xml
Now environment variables can be used for token substitution in the perfmonconfig.xml file.

By defafult system properties are checked.  For example with the variable ${USERNAME}.  First we look for a System property named USERNAME, if one is found the value is returned.  If a system is not found we check for the envirnonment variable named USERNAME.

If you would like to for the environment variables to be checked first, use a token pattern of ${env.USERNAME}..   In this example we first look for an environment variable named USERNAME, if one is not found we look for a system property named USERNAME.