IntervalID now set as a BIGINT (on create only)
After 2.5 years of monitoring many systems we overflowed the INTEGER column primary key of the P4JIntervalData table and the referenced column in the P4JIntervalThreshold table.   

Based on the SQL hoops (dropping indexes and constraints) that would be required for an  automated upgrade we decided not to upgrade existing tables to a BIGINT.  
However, new databases will be created with a BIGINT column instead of a INTEGER Column.   

For those with existing tables a manual update of the column types will be required.