Added context name for JBoss/Wildfly
This is only needed for deploying the perfmon4j web application from within eclipse.  In the distribution package the name of the war file is "perfmon4j.war" however the actual module created in eclipse is "perfmon4j-webapp.jar"  which creates a root context that will not work as a datasource for perfmon4j-reports.