Made isActive() much more efficient
ExternalAppender.isActive() is called by PerfMonTimer.start() .  This means that if a PerfMonTimer is located in a VERY active part of code this method will be called A LOT!  (Actually somewhat paradoxically it only was called when the perfmon4j hadn't been initialized, or the perfmonconfig.xml file had  the global enabled flag set to false... which was rare) 

The prior implementation of this had an || condition, and in a production application enable was always set to false.  This meant that, in production, whenever we would call this method,  we would always enter a synchronization block to check on the size of a map to determine if any remote appenders where attached.

This created a synchronization blocking condition in a high volume production application.

To address this the following changes were made.  
1 - The || was replaced with an &&; and
2 - enabled is now set true on perfmon4j agent initialization when an external port is opened (using the -p javaagent parameter) to allow external appenders to attach.
3 - We now store the active number of remote appenders in an AutomicInteger, so we don't have to wait in a synchronization block to get the count.