Renamed API package
I forgot that when running in JBoss/Wildfly  "org.perfmon4j" classes can only be loaded out of the root classloader when using perfmon4j as a javaagent.   To get around this I took the path of least resistance and simply renamed the package from "org.perfmon4j.agent.api" to "api.org.perfmon4j.agent".  Not as clean, but just as effective.