Feature/sub category splitter for influx appender (#33)
* Disabled testExceptionTrackerOutput_AT

* Updated to allow median calculation

Since we are tracking nano-seconds in this performance test, the mean was almost unusable for comparing one version of perfmon4j to another.  There could be a significant variance with multiple runs of this test with the same perfmon4j version.

* Added test of regex pattern

* New utility

This utility will be used in the Influx appender to allow categories to be grouped together under the same name.  For reporting a subCategory tag will be added to the measurement in InfluxDb

* Added SubCategorySplitter.

The InfluxAppender can now be configured with a RegEx that can be used to group related categories by including a subCategory tag.

* Handle prefix "Interval." or "Snapshot."

With an dynamic appender, like the InfluxAppender, Perfmon4j automatically prepends the category with one of the above prefixes.  This allows users to ignore the prefix when preparing the regex.

* Added SubCategorySplitter

...to LogAnalyticsAppender

* Updated version to 1.6.1-SNAPSHOT

* Fix to ensure multiple patterns can be concatenated