Don't default to log4j logging under quarkus.
Quarkus does not use log4j logging.   When running under Jboss the system property "java.util.logging.manager=org.jboss.logmanager.LogManager" is set, and we have always used that to assume the log4j was being used.  Even though the property is set, that is not the case in Quarkus where java logging is used.

We were