Added testing on scheduling thread traces.
For a long time I thought there were instances where adding a thread trace to the perfmonconfig.xml would not take effect until a system restart.  I tried, without success to create, to create a test scenario to expose this.   I was unable to find anyway for this to occur.  Even though this did not result in any code changes I will be leaving the test in place to provide additional coverage.